+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - CHANGELOG.md file
 - Cookie, LocalStorage and SessionStorage manager
+- Add `html.js`: escapeHtml(), unscapeHtml() and getNodeData()
+
 ### Changed
 - Enhance docs files for better explanation on how to use the boilerplate